+++ R/default_params_doc.R
@@ -23,6 +23,10 @@
 #' i.e. the phenotypic value for homozygotes of the common allele
 #' @param bed_filename name of a \code{PLINK} \code{.bed} file
 #' Use \link{read_plink_bed_file} to read a \code{PLINK} \code{.bed} file.
+#' @param bfile the base filename of the binary files (i.e.
+#' a \code{.bed}, \code{.bim} and \code{.fam} file).
+#' This parameter is named after the \code{PLINK}
+#' \code{--bfile} flag
 #' @param bim_filename name of a \code{PLINK} \code{.bim} file
 #' Use \link{read_plink_bim_file} to read a \code{PLINK} \code{.bim} file.
 #' @param calc_phenotype_function a function that calculate the phenotypes
@@ -91,6 +95,9 @@
 #' @param n_snps the number of SNPs
 #' @param os name of the operating system,
 #' as returned by \link[rappdirs]{app_dir}
+#' @param out the base filename of the output files.
+#' This parameter is named after the \code{PLINK}
+#' \code{--out} flag
 #' @param ped_filename name of a \code{PLINK} \code{.ped} file.
 #' Use \link{read_plink_ped_file} to read a \code{PLINK} \code{.ped} file.
 #' @param ped_table a 'pedigree' table.
@@ -227,6 +234,7 @@ default_params_doc <- function(
   base_output_filename,
   base_phenotype_value,
   bed_filename,
+  bfile,
   bim_filename,
   calc_phenotype_function,
   confidence_interval,
@@ -243,6 +251,7 @@ default_params_doc <- function(
   n_individuals,
   n_snps,
   os,
+  out,
   ped_filename,
   ped_table,
   phenotype,