+++ NEWS.md
@@ -11,7 +11,10 @@ Newest versions at top.
    to install and uninstall the PLINK tutorial data
  * Add `read_plink_lmiss_file` to read PLINK `.lmiss` files
  * Add 'read_plink_imiss_file' to read PLINK `.imiss` files
- * Add 'missing_from_bfiles' to run a PLINK `--missing` analysis from
+ * Add 'read_plink_frq_file' to read PLINK `.frq` files
+ * Add 'missing_from_bfile' to run a PLINK `--missing` analysis from
+   PLINK binary files
+ * Add 'freq_from_bfile' to run a PLINK `--freq` analysis from
    PLINK binary files
 
 ### MINOR IMPROVEMENTS